backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
groq==0.4.0
langchain==0.1.0
langchain-community==0.0.15
langchain-groq==0.0.3
sentence-transformers==2.2.2
chromadb==0.4.22
PyPDF2==3.0.1
pandas==2.1.4
python-dotenv==1.0.0
requests==2.31.0 